requirements.txt
python-telegram-bot~=20.7
python-json-logger~=4.0.0